Refactored table functions. Commented all style code.
