+++ appserver/static/react_native/src/index.jsx
@@ -1,15 +1,16 @@
 import React from 'react'
-import { getUserTheme } from '@splunk/splunk-utils/themes'
 import App from './App'
-// import { SnackbarProvider } from 'notistack'
-import ToastMessages from '@splunk/react-toast-notifications/ToastMessages'
-import { SplunkThemeProvider, mixins, variables /* pick */ } from '@splunk/themes'
-import styled, { createGlobalStyle } from 'styled-components'
-import layout from '@splunk/react-page'
+// import { getUserTheme } from '@splunk/splunk-utils/themes'
+import { SnackbarProvider } from 'notistack'
+// import ToastMessages from '@splunk/react-toast-notifications/ToastMessages'
+// import { SplunkThemeProvider, mixins, variables /* pick */ } from '@splunk/themes'
+// import styled, { createGlobalStyle } from 'styled-components'
+// import layout from '@splunk/react-page'
 // import withSplunkTheme from '@splunk/themes/withSplunkTheme'
 
 // const baseTheme = getTheme({family: 'prisma', colorScheme: 'light', density: 'compact' });
 
+/*
 const GlobalStyles = createGlobalStyle`
   body {
       background-color: ${variables.backgroundColorPage};
@@ -23,6 +24,7 @@ const StyledContainer = styled.div`
     line-height: 200%;
     margin: ${variables.spacing};
 `
+*/
 
 class main extends React.Component {
   /* componentDidMount = () => {
@@ -36,27 +38,30 @@ class main extends React.Component {
   // <SplunkThemeProvider> family="enterprise" colorScheme="dark" density="comfortable">
   render () {
     return (
-      getUserTheme().then(theme => {
+      <SnackbarProvider>
+        <App
+          {...this.props.data}
+          app_context={this.props.utils.getCurrentApp()}
+          utils={this.props.utils}
+          splunk={this.props.splunk.createService({ owner: 'nobody' })}
+          splunk_components={this.props.splunk_components}
+          splunk_theme={this.state?.theme}
+        />
+      </SnackbarProvider>
+      /* getUserTheme().then(theme => {
         layout(
           <SplunkThemeProvider>
             <GlobalStyles />
             <StyledContainer>
-              <App
-                {...this.props.data}
-                app_context={this.props.utils.getCurrentApp()}
-                utils={this.props.utils}
-                splunk={this.props.splunk.createService({ owner: 'nobody' })}
-                splunk_components={this.props.splunk_components}
-                splunk_theme={this.state?.theme}
-              />
-              <ToastMessages position='top-center' />
+
+            <ToastMessages position='top-center' />
             </StyledContainer>
           </SplunkThemeProvider>,
           {
             theme
           }
-        )
-      })
+      //  )
+      // }) */
     )
   }
 }
